+++ docs/src/Fulma/Components/Panel.fs
@@ -37,6 +37,67 @@ let iconInteractive () =
                                       Button.IsFullWidth ]
                                     [ str "Reset" ] ] ] ] ]
 
+let colorInteractive () =
+    Columns.columns [ ]
+        [ Column.column [ ]
+            [ Panel.panel [ Panel.Color IsPrimary ]
+                [ Panel.heading [ ] [ str "Repositories"]
+                  Panel.block [ ]
+                    [ Control.div [ Control.HasIconLeft ]
+                        [ Input.text [ Input.Size IsSmall
+                                       Input.Placeholder "Search" ]
+                          Icon.icon [ Icon.Size IsSmall
+                                      Icon.IsLeft ]
+                                    [ i [ ClassName "fa fa-search" ] [ ] ] ] ]
+                  Panel.tabs [ ]
+                    [ Panel.tab [ ] [ str "All" ]
+                      Panel.tab [ Panel.Tab.IsActive true ] [ str "Fable" ]
+                      Panel.tab [ ] [ str "Elmish" ]
+                      Panel.tab [ ] [ str "Bulma" ] ]
+                  Panel.block [ Panel.Block.IsActive true ]
+                    [ Panel.icon [ ] [ i [ ClassName "fa fa-book" ] [ ] ]
+                      str "Bulma" ]
+                  Panel.block [ ]
+                    [ Panel.icon [ ] [ i [ ClassName "fa fa-code-fork" ] [ ] ]
+                      str "Fable" ]
+                  Panel.checkbox [ ]
+                    [ input [ Type "checkbox" ]
+                      str "I am a checkbox" ]
+                  Panel.block [ ]
+                    [ Button.button [ Button.Color IsPrimary
+                                      Button.IsOutlined
+                                      Button.IsFullWidth ]
+                                    [ str "Reset" ] ] ] ]
+          Column.column [ ]
+            [ Panel.panel [ Panel.Color IsDanger ]
+                [ Panel.heading [ ] [ str "Repositories"]
+                  Panel.block [ ]
+                    [ Control.div [ Control.HasIconLeft ]
+                        [ Input.text [ Input.Size IsSmall
+                                       Input.Placeholder "Search" ]
+                          Icon.icon [ Icon.Size IsSmall
+                                      Icon.IsLeft ]
+                                    [ i [ ClassName "fa fa-search" ] [ ] ] ] ]
+                  Panel.tabs [ ]
+                    [ Panel.tab [ ] [ str "All" ]
+                      Panel.tab [ Panel.Tab.IsActive true ] [ str "Fable" ]
+                      Panel.tab [ ] [ str "Elmish" ]
+                      Panel.tab [ ] [ str "Bulma" ] ]
+                  Panel.block [ Panel.Block.IsActive true ]
+                    [ Panel.icon [ ] [ i [ ClassName "fa fa-book" ] [ ] ]
+                      str "Bulma" ]
+                  Panel.block [ ]
+                    [ Panel.icon [ ] [ i [ ClassName "fa fa-code-fork" ] [ ] ]
+                      str "Fable" ]
+                  Panel.checkbox [ ]
+                    [ input [ Type "checkbox" ]
+                      str "I am a checkbox" ]
+                  Panel.block [ ]
+                    [ Button.button [ Button.Color IsPrimary
+                                      Button.IsOutlined
+                                      Button.IsFullWidth ]
+                                    [ str "Reset" ] ] ] ] ]
+
 
 let view =
     Render.docPage [ Render.contentFromMarkdown
@@ -49,4 +110,7 @@ A composable **panel**, for compact controls
                         """
                      Render.docSection
                         ""
-                        (Widgets.Showcase.view iconInteractive (Render.includeCode __LINE__ __SOURCE_FILE__)) ]
+                        (Widgets.Showcase.view iconInteractive (Render.includeCode __LINE__ __SOURCE_FILE__))
+                     Render.docSection
+                        "### Colors"
+                        (Widgets.Showcase.view colorInteractive (Render.includeCode __LINE__ __SOURCE_FILE__)) ]